+++ CHANGELOG.md
@@ -8,10 +8,36 @@ All versions prior to 0.0.9 are untracked.
 
 ## [Unreleased]
 
+## [2.9.0]
+
+* `pip-audit` now supports [PEP 751](https://peps.python.org/pep-0751/)
+  lockfiles. These lockfiles can be audited in "project" mode by
+  passing `--locked` to `pip-audit`
+  ([#888](https://github.com/pypa/pip-audit/pull/888))
+
+## [2.8.0]
+
 ### Added
 
 * `pip-audit` now allows some CLI flags to be configured via environment
   variables ([#755](https://github.com/pypa/pip-audit/pull/755))
+
+### Changed
+
+* The default cache locations on macOS and Linux now respect each platform's
+  caching directory idioms (e.g. XDG)
+  ([#814](https://github.com/pypa/pip-audit/pull/814))
+
+* The minimum version of Python is now 3.9
+  ([#846](https://github.com/pypa/pip-audit/pull/846))
+
+### Fixed
+
+* Auditing a fully-pinned requirements file with `--disable-pip` now allows for
+  duplicates, so long as the duplicates don't have conflicting specifier sets
+  ([#749](https://github.com/pypa/pip-audit/pull/749))
+* Fixed two sources of unnecessary resource leaks when doing file I/O
+  ([#878](https://github.com/pypa/pip-audit/pull/878))
 
 ## [2.7.3]
 
@@ -599,7 +625,9 @@ All versions prior to 0.0.9 are untracked.
   dependency errors ([#146](https://github.com/pypa/pip-audit/pull/146))
 
 <!-- Release URLs -->
-[Unreleased]: https://github.com/pypa/pip-audit/compare/v2.7.3...HEAD
+[Unreleased]: https://github.com/pypa/pip-audit/compare/v2.9.0...HEAD
+[2.9.0]: https://github.com/pypa/pip-audit/compare/v2.8.0...v2.9.0
+[2.8.0]: https://github.com/pypa/pip-audit/compare/v2.7.3...v2.8.0
 [2.7.3]: https://github.com/pypa/pip-audit/compare/v2.7.2...v2.7.3
 [2.7.2]: https://github.com/pypa/pip-audit/compare/v2.7.1...v2.7.2
 [2.7.1]: https://github.com/pypa/pip-audit/compare/v2.7.0...v2.7.1